add cap 3 task
